[STY] Use a shortcut to context and thread context.  Type adjustements.

+++ blosc/context.h
@@ -103,7 +103,7 @@ struct thread_context {
   uint8_t* tmp2;
   uint8_t* tmp3;
   uint8_t* tmp4;
-  size_t tmpblocksize; /* keep track of how big the temporary buffers are */
+  int32_t tmpblocksize; /* keep track of how big the temporary buffers are */
 #if defined(HAVE_ZSTD)
   /* The contexts for ZSTD */
   ZSTD_CCtx* zstd_cctx;
